moved syncsettings to generic settings object for more extensibility in future
